backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
pydantic>=2.7.0
python-multipart>=0.0.9
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
selenium>=4.21.0
google-generativeai>=0.7.2
